step-16 Moved authentication views into a separate folder under the view directory. Created a new dialog to confirm whether the user wanted to reset their application preferences. Included more specifications for the dark and light themes in the main.dart file to make the text buttons looked consistent with the rest of the application.
