+++ lib/main.dart
@@ -7,12 +7,12 @@ import 'package:mobile_motivation/services/auth/authBloc/auth_bloc.dart';
 import 'package:mobile_motivation/services/auth/firebase_auth_provider.dart';
 import 'package:mobile_motivation/services/system/cubit/preferences_cubit.dart';
 import 'package:mobile_motivation/services/system/preferences_service.dart';
-import 'package:mobile_motivation/views/forgot_password_view.dart';
-import 'package:mobile_motivation/views/login_view.dart';
+import 'package:mobile_motivation/views/authentication_ui/forgot_password_view.dart';
+import 'package:mobile_motivation/views/authentication_ui/login_view.dart';
 import 'package:mobile_motivation/views/main_ui/home_view.dart';
 import 'package:mobile_motivation/views/main_ui/settings_view.dart';
-import 'package:mobile_motivation/views/register_view.dart';
-import 'package:mobile_motivation/views/verify_email_view.dart';
+import 'package:mobile_motivation/views/authentication_ui/register_view.dart';
+import 'package:mobile_motivation/views/authentication_ui/verify_email_view.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 void main() async {
@@ -61,21 +61,43 @@ class MainApp extends StatelessWidget {
           // to the Preferences bloc.
           return BlocProvider(
             create: (_) => blocSnapshot.data!,
+            // We need a BlocBuilder to build the MaterialApp and giving
+            // control to the Preferences bloc such that it can provide
+            // the correct theme mode via the Preferences data.
             child: BlocBuilder<PreferencesCubit, Preferences>(
               builder: (context, preferences) => MaterialApp(
                 title: 'Flutter Demo',
                 // Change theme data of light and dark such that each
-                // utilize the deep purple color for the background of app bars.
+                // utilize the deep purple color for the background of app bars
+                // and the text for the TextButtons are different.
                 theme: ThemeData.light().copyWith(
                   appBarTheme: const AppBarTheme(
                     backgroundColor: Colors.deepPurple,
+                  ),
+                  textButtonTheme: TextButtonThemeData(
+                    style: TextButton.styleFrom(
+                      textStyle: const TextStyle(
+                        color: Colors.black,
+                      ),
+                    ),
                   ),
                 ),
                 darkTheme: ThemeData.dark().copyWith(
                   appBarTheme: const AppBarTheme(
                     backgroundColor: Colors.deepPurple,
                   ),
+                  textButtonTheme: TextButtonThemeData(
+                    style: TextButton.styleFrom(
+                      textStyle: const TextStyle(
+                        color: Colors.white,
+                      ),
+                    ),
+                  ),
                 ),
+                // PreferencesBloc is utilized here to control the themeMode
+                // of the application by using its preferences member
+                // depending on whether the app was just opened or if the user
+                // changed the theme in the settings.
                 themeMode: preferences.themeMode,
                 // BlocProvider provides the AuthBloc instance and has a child widget
                 // which is the HomePage (since the widget must return a BlocBuilder).
