+++ lib/views/authentication_ui/register_view.dart
