+++ lib/views/authentication_ui/verify_email_view.dart
